Fix not all floats fit a decimal

+++ modbus-notmodbus/TelemetryPoint.cs
@@ -5,6 +5,6 @@ namespace modbus_notmodbus
     public class TelemetryPoint
     {
         public string iotHubDeviceId;
-        public Dictionary<string, decimal> analogInput;
+        public Dictionary<string, float> analogInput;
     }
 }
